Consumer automobile loans dataset

A dataset containing the performance records of 58,118 consumer automobile loans with original terms of 72 to 73 months. The data was obtained from four publicly traded asset-backed securities (ABS) bonds. This dataset is subject to left-truncation (loans must survive long enough to be included in the ABS trust) and right-censoring (the trust may be unwound before all loans terminate), and it features competing risks of default and prepayment.

Usage

aloans

Format

aloans

A data frame with 58,118 rows and 12 columns:

risk.cat

Credit risk band based on the loan's Annual Percentage Rate (APR). Categories include "super_prime" (0-5%), "prime" (5-10%), "near_prime" (10-15%), "subprime" (15-20%), and "deep_subprime" (20%+).

Z

Time-to-event, representing the loan termination age in months.

Y

Left-truncation time, representing the loan age in months when the ABS trust began making payments to investors.

C

Right-censoring indicator (1 = right-censored, 0 = exact termination observed).

D

Default indicator for competing risks (1 = default, 0 = prepayment).

bond

Categorical variable specifying the source ABS bond: CARMX (CarMax Auto Owner Trust 2017-2), AART (Ally Auto Receivables Trust 2017-3), SDART (Santander Drive Auto Receivables Trust 2017-2), or DRIVE (Drive Auto Receivables Trust 2017-1).

orig.apr

Obligor annual percentage rate

orig.term

Original loan term

orig.loan.amt

Original loan amount

cur.age

Loan age as of the start of the asset-backed security

cur.balance

Outstanding loan balance as of the start of the asset-backed security

calc.pmt

Calculated monthly payment

Source

Data was originally compiled from the the Electronic Data Gathering, Analysis, and Retrieval (EDGAR) system of the Securities and Exchange Comission (SEC).. The replication data repository is available at https://github.com/jackson-lautier/credit-risk-convergence/.

References

Lautier, J. P., Pozdnyakov, V., & Yan, J. (2024). On the convergence of credit risk in current consumer automobile loans. Journal of the Royal Statistical Society Series A: Statistics in Society, qnae137. doi:10.1093/jrsssa/qnae137 .